moved secret key and debug from common to ex settings
If you deploy in an environment where you don't have read or write access to
the secret key location, it's impossible to override the secret key without
getting access denied errors. So the secret key logic makes more sense to live
in the example settings where it can be taken out on a per-deployment basis.

DEBUG is set to True explicitly in the example settings now, to make it an
explicit choice where to set it rather than having it default to true in common
settings.