fixed relative import
the deploy settings file is in a different package and wasn't properly
importing common settings from where it is.

This assumes the default way to run the deployment in production is to point
the WSGI server at the settings file with "project.deploy.settings", rather
than e.g. copying it and overwriting project/settings.py